requirements.txt
setuptools
wheel
pyserial
Homie4
